// src/app/model/user.ts
export interface User {
  id?: number;
  email: string;
  password: string;
  firstName?: string;
  lastName?: string;
  role?: string;
}
export interface LoginRequest {
  email: string;
  password: string;
}

export interface RegisterRequest {
  email: string;
  password: string;
  firstName?: string;
  lastName?: string;
}

export interface AuthResponse {
  user: User;
  token: string;
}